+++ all/cancel/index.js
@@ -1,0 +1,6 @@
+export class Cancel extends Error {
+	constructor(config) {
+		super(config?.message ?? 'Cancel')
+		this.returns = config?.returns
+	}
+}