+++ all/promise_timeout/index.d.ts
@@ -1,5 +1,6 @@
+import type { cancel_Promise } from '../promise/index.js'
 export declare function promise_timeout<O>(
 	promise:(()=>Promise<O>)|Promise<O>,
 	ms:number,
 	error?:Error
-):Promise<O>&{ cancel:(val:O)=>void }
+):cancel_Promise<O, (val?:O)=>Promise<O>>