+++ all/rmemo__wait/index.d.ts
@@ -1,3 +1,4 @@
+import { cancel_Promise } from '../promise/index.js'
 import type { memo_T, rmemo_T, rmemo_val_T } from '../rmemo/index.js'
 export declare function rmemo__wait<
 	_rmemo_T extends rmemo_T<unknown> = rmemo_T<unknown>
@@ -6,8 +7,12 @@ export declare function rmemo__wait<
 	condition_fn:(val:rmemo_val_T<_rmemo_T>)=>unknown,
 	timeout?:number,
 	error?:Error
-):Promise<rmemo_val_T<_rmemo_T>>&{
+):rmemo__wait_ret_T<rmemo_val_T<_rmemo_T>>
+export type rmemo__wait_ret_T<R> =
+	cancel_Promise<
+		R,
+		(val?:R)=>Promise<R>
+	>&{
 	// prevent early usGC
 	m:memo_T<unknown>
-	cancel:(val:rmemo_val_T<_rmemo_T>)=>void
 }