+++ all/sleep/index.d.ts
@@ -1,1 +1,3 @@
-export declare function sleep(ms:number):Promise<number|NodeJS.Timeout>
+export declare function sleep(
+	ms:number
+):Promise<number>