+++ all/waitfor/index.js
@@ -1,31 +1,61 @@
+import { Cancel } from '../cancel/index.js'
 import { promise_timeout } from '../promise_timeout/index.js'
 import { sleep } from '../sleep/index.js'
 /**
  * @param {()=>Promise<unknown>}fn
- * @param {unknown}timeout
- * @param {unknown}[period]
- * @returns {Promise<void>}
+ * @param {number}timeout
+ * @param {number|(()=>Promise<unknown>)}[period]
+ * @returns {Promise<unknown>}
  */
-export function waitfor(fn, timeout, period = 0) {
-	let cancel
+export function waitfor(
+	fn,
+	timeout,
+	period
+) {
+	let rv
+	let cancel_arg_a
 	let promise = new Promise((resolve, reject)=>
 		promise_timeout(async ()=>{
-			let rv
-			for (; !cancel;) {
-				rv = await fn()
-				if (rv) return rv
-				await sleep(period)
+			for (; !cancel_arg_a;) {
+				let _rv = await fn()
+				rv = cancel_arg_a?.length ? cancel_arg_a[0] : _rv
+				if (rv || cancel_arg_a) return rv
+				if (typeof period === 'function') {
+					await period(promise)
+				} else {
+					await sleep(period ?? 0)
+				}
 			}
 			return rv
 		}, timeout)
 			.then(resolve)
 			.catch(err=>{
-				cancel = 1
+				cancel_arg_a = []
 				reject(err)
 			}))
-	promise.cancel = ()=>{
-		cancel = 1
+	promise.cancel = (...arg_a)=>{
+		cancel_arg_a = arg_a
 		return promise
 	}
 	return promise
+}
+/**
+ * @param {number}ms
+ * @param {()=>boolean}should_cancel_
+ * @param {Cancel_config_T}[Cancel_config]
+ * @returns {(promise:Promise<unknown>&{cancel():unknown})=>Promise<number>}
+ * @private
+ */
+export function cancel__period_(
+	ms,
+	should_cancel_,
+	Cancel_config
+) {
+	return promise=>{
+		if (should_cancel_()) {
+			promise.cancel()
+			throw new Cancel(Cancel_config)
+		}
+		return sleep(ms)
+	}
 }